refactor(@schematics/angular): migrate typescript third-party dependency to node_modules
Relocates the typescript third-party dependency from `third_party/github.com/Microsoft/TypeScript` to `third_party/typescript`.

This update changes the implementation to source the required TypeScript files directly from `node_modules` via a Bazel `genrule` instead of keeping them checked into the repository.
