+++ packages/schematics/angular/app-shell/index.ts
@@ -15,7 +15,7 @@ import {
   schematic,
 } from '@angular-devkit/schematics';
 import { dirname, join } from 'node:path/posix';
-import ts from '../third_party/github.com/Microsoft/TypeScript/lib/typescript';
+import ts from '../third_party/typescript';
 import {
   findNode,
   findNodes,
