+++ packages/schematics/angular/refactor/jasmine-vitest/transformers/jasmine-matcher.ts
@@ -14,7 +14,7 @@
  * `toHaveBeenCalledOnceWith` and `arrayWithExactContents`.
  */
 
-import ts from '../../../third_party/github.com/Microsoft/TypeScript/lib/typescript';
+import ts from '../../../third_party/typescript';
 import {
   addVitestValueImport,
   createExpectCallExpression,
