+++ packages/schematics/angular/refactor/jasmine-vitest/transformers/jasmine-type.ts
@@ -13,7 +13,7 @@
  * necessary `vitest` imports are added to the file.
  */
 
-import ts from '../../../third_party/github.com/Microsoft/TypeScript/lib/typescript';
+import ts from '../../../third_party/typescript';
 import { addVitestTypeImport } from '../utils/ast-helpers';
 import { RefactorContext } from '../utils/refactor-context';
