+++ packages/schematics/angular/refactor/jasmine-vitest/utils/ast-helpers.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import ts from '../../../third_party/github.com/Microsoft/TypeScript/lib/typescript';
+import ts from '../../../third_party/typescript';
 
 export function addVitestValueImport(imports: Set<string>, importName: string): void {
   imports.add(importName);
