+++ packages/schematics/angular/refactor/jasmine-vitest/utils/comment-helpers.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import ts from '../../../third_party/github.com/Microsoft/TypeScript/lib/typescript';
+import ts from '../../../third_party/typescript';
 import { TODO_NOTES, TodoCategory, TodoContextMap } from './todo-notes';
 
 // A helper type to find all `TodoCategory` keys that do not require a context.
