+++ packages/schematics/angular/refactor/jasmine-vitest/utils/comment-helpers_spec.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import ts from '../../../third_party/github.com/Microsoft/TypeScript/lib/typescript';
+import ts from '../../../third_party/typescript';
 import { addTodoComment } from './comment-helpers';
 
 describe('addTodoComment', () => {
