+++ packages/schematics/angular/utility/add-declaration-to-ng-module.ts
@@ -7,7 +7,7 @@
  */
 
 import { Rule, Tree, strings } from '@angular-devkit/schematics';
-import * as ts from '../third_party/github.com/Microsoft/TypeScript/lib/typescript';
+import * as ts from '../third_party/typescript';
 import { addDeclarationToModule, addSymbolToNgModuleMetadata } from './ast-utils';
 import { InsertChange } from './change';
 import { buildRelativePath } from './find-module';
