+++ packages/schematics/angular/utility/ast-utils_spec.ts
@@ -7,7 +7,7 @@
  */
 
 import { HostTree } from '@angular-devkit/schematics';
-import * as ts from '../third_party/github.com/Microsoft/TypeScript/lib/typescript';
+import * as ts from '../third_party/typescript';
 import { Change, InsertChange } from '../utility/change';
 import {
   addDeclarationToModule,
